Use Cleo commands instead of scripts
