+++ pyproject.toml
@@ -9,8 +9,6 @@ authors = [
 
 [tool.poetry.scripts]
 mozci = 'mozci.console.application:cli'
-classify_batch_execution = "mozci.util.classify_batch_execution:main"
-evaluate_batch_execution = "mozci.util.evaluate_batch_execution:main"
 
 [tool.poetry.dependencies]
 python = ">=3.7,<4"
